requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
reportlab==4.0.4
flask-login==0.6.2
flask-wtf==1.2.1
wtforms==3.1.2
email-validator==2.1.1
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
gunicorn==22.0.0
Werkzeug==2.3.7
  


